Add a solve function for problem 5
Defines a function that takes an immutable reference to a 2d matrix of
integers and returns a maximum value of sum of elements in a submatrix.
The functions uses map() method to create a vector of sums of rows, then
uses max() method to find the maximum sum of submatrix.

Close: #10