Add a solve function for problem 7
Defines a funtion that takes a four digit number and split it into
digits. It sorts the digits in ascending order and place the two
smallest digits in the tens place and the two largest digits in the
onces place to obtain a pair of two digit numbers with the minimum
possible sum.

Add problem7.rs

ommit file