Add a solve function for problem3 (#7)
Define a function that takes an immutable reference to a vector of
integers 'v1' as an input parameter. The function returns a new vector
of integers 'v2', which satisfies a following condition: v2[i] =
v1[v1[i]]. In order to satisfying the condition, The function iterates
over the input vector 'v1' and pushes the element at index 'v1[i]' to
the output vector 'v2'. Finally, it can return the output vector 'v2'.