Add a solve function for problem 4 (#9)
Defines a function that takes an immutable reference to a vector of
integers as an input parameter and returns a new vector of integers
containing a prefix sum of the input vector. The function defines a new
integer variable to store the prefix sum of the input, and updates
through iteration over the input vector. It also pushes the prefix sum
to a new vector while iteration, and returns the new vector as a result.

Close: #8