Add a solve function for problem 9
Defines a function that returns a boolean array 'result', where
result[i] is true if the ith kid will have the greatest number of
candies, or false otherwise. The function works by finding the maximum
by using a max() method, and then comparing each element with the
maximum while adding the extra candies. It returns the boolean array
'result' by using a map(), and collect() method.