Add a solve function for problem 8 (#17)
Defines a function that takes a array of integers and calculates the
number of good pairs in the array. A good pair is a pair of integers in
the array that have the same value. The function defines a hashmap to
keep the number of times each integer appears in the array. It then
calculates the combinatorial of the number of times each integer appears
(nC2) and sums the results to get the total number of good pairs.

Close: #16