Add a solve function for problem 11 (#23)
Defines a function that creates a sorted version of the input vector and
iterates over the sorted vector while tracking the first occurrence of
each value. The function store these values in a HashMap, using the
number as the key and its index as the value. Finally, the function
iterates over the original vector and retrieve the value from the
HashMap. It stores in a result vector and returns it as the output.

Close: #22